+++ app/src/code/buffer_location_tests.rs
@@ -756,203 +756,3 @@ fn handle_buffer_updated_push_multiple_edits_in_batch() {
         assert_eq!(content(&app, file_id), "xxx bbb zzz");
     })
 }
-
-// ── Pending edit batch: discard on server push ───────────────────
-
-#[test]
-fn pending_batch_discarded_on_server_push_with_conflict() {
-    App::test((), |mut app| async move {
-        init_app(&mut app);
-        app.add_singleton_model(GlobalBufferModel::new);
-
-        let host_id = test_host_id();
-        let path = test_path();
-
-        // Seed a remote buffer at server_version=1, client_version=0.
-        let _buffer_state = gbm(&app).update(&mut app, |gbm, ctx| {
-            gbm.seed_remote_buffer_for_test(host_id.clone(), path.clone(), "hello", 1, ctx)
-        });
-        let file_id = _buffer_state.file_id;
-
-        // Simulate client edits that haven't been flushed yet.
-        let client_cv = ContentVersion::new();
-        gbm(&app).update(&mut app, |gbm, _ctx| {
-            gbm.insert_pending_batch_for_test(
-                file_id,
-                1, // expected_server_version
-                vec![text_edit(6, 6, " world")],
-                client_cv,
-            );
-        });
-
-        // Verify the batch exists.
-        let handle = gbm(&app);
-        app.read(|ctx| {
-            assert!(handle.as_ref(ctx).has_pending_batch_for_test(file_id));
-            assert_eq!(
-                handle
-                    .as_ref(ctx)
-                    .pending_batch_edit_count_for_test(file_id),
-                1
-            );
-        });
-
-        // Server push arrives. Since client_cv != 0, this triggers a conflict
-        // and the batch should be discarded.
-        gbm(&app).update(&mut app, |gbm, ctx| {
-            gbm.handle_buffer_updated_push(
-                &host_id,
-                path.as_str(),
-                2, // new_server_version
-                0, // expected_client_version (server doesn't know about our edits)
-                &[char_edit(6, 6, " push")],
-                ctx,
-            );
-        });
-
-        // Batch should be discarded.
-        let handle = gbm(&app);
-        app.read(|ctx| {
-            assert!(
-                !handle.as_ref(ctx).has_pending_batch_for_test(file_id),
-                "Pending batch should be discarded on server push"
-            );
-        });
-
-        // Content should be unchanged (conflict path, push not applied).
-        assert_eq!(content(&app, file_id), "hello");
-    })
-}
-
-#[test]
-fn pending_batch_discarded_on_conflict_detected() {
-    App::test((), |mut app| async move {
-        init_app(&mut app);
-        app.add_singleton_model(GlobalBufferModel::new);
-
-        let host_id = test_host_id();
-        let path = test_path();
-
-        let _buffer_state = gbm(&app).update(&mut app, |gbm, ctx| {
-            gbm.seed_remote_buffer_for_test(host_id.clone(), path.clone(), "hello", 1, ctx)
-        });
-        let file_id = _buffer_state.file_id;
-
-        // Insert a pending batch.
-        let client_cv = ContentVersion::new();
-        gbm(&app).update(&mut app, |gbm, _ctx| {
-            gbm.insert_pending_batch_for_test(
-                file_id,
-                1,
-                vec![text_edit(6, 6, " edit")],
-                client_cv,
-            );
-        });
-
-        let handle = gbm(&app);
-        app.read(|ctx| {
-            assert!(handle.as_ref(ctx).has_pending_batch_for_test(file_id));
-        });
-
-        // BufferConflictDetected arrives.
-        gbm(&app).update(&mut app, |gbm, ctx| {
-            gbm.handle_buffer_conflict_detected(&host_id, path.as_str(), ctx);
-        });
-
-        // Batch should be discarded.
-        let handle = gbm(&app);
-        app.read(|ctx| {
-            assert!(
-                !handle.as_ref(ctx).has_pending_batch_for_test(file_id),
-                "Pending batch should be discarded on conflict detected"
-            );
-        });
-    })
-}
-
-#[test]
-fn server_push_accepted_without_pending_batch() {
-    App::test((), |mut app| async move {
-        init_app(&mut app);
-        app.add_singleton_model(GlobalBufferModel::new);
-
-        let host_id = test_host_id();
-        let path = test_path();
-
-        let _buffer_state = gbm(&app).update(&mut app, |gbm, ctx| {
-            gbm.seed_remote_buffer_for_test(host_id.clone(), path.clone(), "hello", 1, ctx)
-        });
-        let file_id = _buffer_state.file_id;
-
-        // No pending batch — clean push should be accepted.
-        let handle = gbm(&app);
-        app.read(|ctx| {
-            assert!(!handle.as_ref(ctx).has_pending_batch_for_test(file_id));
-        });
-
-        gbm(&app).update(&mut app, |gbm, ctx| {
-            gbm.handle_buffer_updated_push(
-                &host_id,
-                path.as_str(),
-                2,
-                0, // matches client_version=0
-                &[char_edit(6, 6, " world")],
-                ctx,
-            );
-        });
-
-        assert_eq!(content(&app, file_id), "hello world");
-
-        // Clock should be updated.
-        let handle = gbm(&app);
-        app.read(|ctx| {
-            let clock = handle
-                .as_ref(ctx)
-                .sync_clock_for_remote_test(file_id)
-                .unwrap();
-            assert_eq!(clock.server_version, ContentVersion::from_raw(2));
-            assert_eq!(clock.client_version, ContentVersion::from_raw(0));
-        });
-    })
-}
-
-#[test]
-fn pending_batch_bumps_client_version_immediately() {
-    App::test((), |mut app| async move {
-        init_app(&mut app);
-        app.add_singleton_model(GlobalBufferModel::new);
-
-        let host_id = test_host_id();
-        let path = test_path();
-
-        let _buffer_state = gbm(&app).update(&mut app, |gbm, ctx| {
-            gbm.seed_remote_buffer_for_test(host_id.clone(), path.clone(), "hello", 1, ctx)
-        });
-        let file_id = _buffer_state.file_id;
-
-        // Insert a batch — this simulates what the ContentChanged handler does:
-        // sync_clock.client_version is bumped immediately.
-        let client_cv = ContentVersion::new();
-        gbm(&app).update(&mut app, |gbm, _ctx| {
-            gbm.insert_pending_batch_for_test(
-                file_id,
-                1,
-                vec![text_edit(6, 6, " edit")],
-                client_cv,
-            );
-        });
-
-        // The sync clock's client_version should already reflect the edit,
-        // even though the batch hasn't been flushed.
-        let handle = gbm(&app);
-        app.read(|ctx| {
-            let clock = handle
-                .as_ref(ctx)
-                .sync_clock_for_remote_test(file_id)
-                .unwrap();
-            assert_eq!(clock.client_version, client_cv);
-            // server_version unchanged
-            assert_eq!(clock.server_version, ContentVersion::from_raw(1));
-        });
-    })
-}
